Calculator/calculator/src/main/java/ca/vanier/CalculatorResponse.java
package ca.vanier;

public class CalculatorResponse {

    String operation, detailMessage, Status;
    double result, value1, value2;

    public String getOperation() {
        return operation;
    }

    public void setOperation(String operation) {
        this.operation = operation;
    }

    public double getValue1() {
        return value1;
    }

    public void setValue1(double value1) {
        this.value1 = value1;
    }

    public double getValue2() {
        return value2;
    }

    public void setValue2(double value2) {
        this.value2 = value2;
    }

    public double getResult() {
        return result;
    }

    public void setResult(double result) {
        this.result = result;
    }

    public String getDetailMessage() {
        return detailMessage;
    }

    public void setDetailMessage(String detailMessage) {
        this.detailMessage = detailMessage;
    }

    public String getStatus() {
        return Status;
    }

    public void setStatus(String status) {
        Status = status;
    }

    @Override
    public String toString() {
        return "CalculatorResponse [operation=" + operation + ", value1=" + value1 + ", value2=" + value2 + ", result="
                + result + ", detailMessage=" + detailMessage + ", Status=" + Status + "]";
    }

}
